refactor(api): move key permission requirements into their owning routes
