+++ svc/api/routes/v2_keys_reroll_key/handler.go
@@ -26,7 +26,6 @@ import (
 	"github.com/unkeyed/unkey/pkg/urn"
 	"github.com/unkeyed/unkey/pkg/zen"
 	"github.com/unkeyed/unkey/svc/api/internal/auditactor"
-	"github.com/unkeyed/unkey/svc/api/internal/keyperms"
 )
 
 type (
@@ -401,14 +400,9 @@ func (h *Handler) RerollKey(
 // routes pass their product capability directly instead.
 func rerollPermissionQuery(key db.FindLiveKeyByIDRow) rbac.PermissionQuery {
 	keyData := db.ToKeyData(key)
-	scope := keyperms.Scope{
-		WorkspaceID: key.WorkspaceID,
-		KeyspaceID:  key.KeyAuthID,
-		APIID:       key.Api.ID,
-	}
 
 	checks := rbac.Or(
-		keyperms.CreateKey(scope),
+		CreateKeyPermissions(key.Api.ID),
 		rbac.U(
 			urn.New().Workspace(key.WorkspaceID).Keyspace(key.KeyAuthID),
 			permissions.CreateKey{},
@@ -419,7 +413,7 @@ func rerollPermissionQuery(key db.FindLiveKeyByIDRow) rbac.PermissionQuery {
 		checks = rbac.And(
 			checks,
 			rbac.Or(
-				keyperms.EncryptKey(scope),
+				EncryptKeyPermissions(key.Api.ID),
 				rbac.U(
 					urn.New().Workspace(key.WorkspaceID).Keyspace(key.KeyAuthID).Key("*"),
 					permissions.EncryptKey{},
